Remove Timezone Dependency
Formerly, the initial record pull was based on the current time of the
server minus five minutes.  This caused timing issues, specifically in
the case where the database server and dbmonitor server had different
timezones set.  The initial record is now grabbed based on the
timestamp of the last record insertted into the table.  If there are
no records in the database 0000-01-01 00:00:00 is used as the initial
timestamp.